requirements.txt
# Core ML and AI libraries
openai-whisper>=20231117
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
torch>=1.13.0
torchaudio>=0.13.0
transformers>=4.21.0

# Web framework and API
flask==2.3.3
flask-cors==4.0.0

# Data processing and utilities
numpy>=1.21.0
pandas>=1.5.0
tqdm>=4.64.0
python-dotenv>=1.0.0

# File handling and JSON
ujson>=5.7.0
pathlib2>=2.3.7

# Logging and configuration
pyyaml>=6.0
colorlog>=6.7.0

# Audio processing (additional support)
librosa>=0.9.2
soundfile>=0.12.1

# Development and testing
pytest>=7.2.0
black>=22.0.0
flake8>=5.0.0